refactor: rename dir name
