+++ vite-plugins/inline-js.ts
